sprite-puppet-animator/backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
rembg==2.0.53
numpy==1.26.3

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# GIF & Animation
imageio==2.33.1
imageio-ffmpeg==0.4.9

# CORS
starlette==0.35.1
